Three dc3 source lanes were dispatched on counts derived from target_symbol_map
.json's residency. The lane replaced the instrument first, and that is what made
it work: orig/373307D9/ham_xbox_r.map is the SHIPPED MSVC linker map and prints
one line per symbol, with fold sets sharing an address -- 118,000 names against
target_symbol_map.json's 69,132. It states a fold outright instead of leaving
the loser invisible. Banked as tools/mapidx.py + tools/reclass.py. Re-read
through it, different_function is 57 of 86 pairs "retail shipped both bodies"
and only 2 folds, i.e. a real source lane rather than a naming artefact.

Verdicts, one per lane:

template_sibling is a MakeString argument-type lane and free to fix -- 9 of 12
pairs. MakeString<T> mangles T, the map says which T's fold (void* = unsigned =
ulong at 823b2008, int = enum at 82610090, char* = const char* at 823314d8), and
every site already materialises a one-word stack temp and passes its address, so
a Symbol temp, a const char* temp and an unsigned temp are the same word in the
same slot. 25 -> 8 sites, 17 -> 5 functions.

local_static_moved_fn is NOT a source lane. Refuted: all 50 solo functions are
placeholder-named 32-byte guard-clear thunks (1,600 B), and 49 of the 54 charged
retail-side guards are already defined by our own object in the same TU. objdiff
pairs interchangeable anonymous thunks by position and our COFF has no symbol
for them either, so symbols.txt cannot fix it -- the repair belongs in objdiff's
matcher. Two of the five exceptions are real: InviteAcceptedMsg and
PartyMembersChangedMsg are genuinely absent from our PlatformMgr_Xbox.obj.

different_function is real, and roughly 60 pairs are sixty separate readings --
about 30 unrelated-class destructors and 12 class-scoped operator deletes. It is
tempting to call them ICF folds; the map says otherwise (??3FlowNode at
823eb968 against ??3PropertyEventProvider at 823ebcc8, same object). We really
do call the wrong pooled delete, one inlined-lifetime site at a time, with no
shared key. The tractable sub-shape -- pairs whose manglings differ in something
the mangling NAMES -- is where all 20 gains came from, and it is now exhausted
in dc3.

What did not work, kept because the measurements are the point:

  * Hoisting a const view out of a loop cost 128 B at `none` (GetLabelTotalAlpha
    100 -> 73.28): the original reloads the ObjPtr through the loop condition.
    The same local declared INSIDE the loop body is byte-free. Same intent, two
    placements, one a regression.
  * MILO_LOG -> MILO_NOTIFY on HamDirector cost 932 B at `none` and was
    reverted. It reads exactly like the five MILO_WARN -> MILO_NOTIFY sites that
    landed clean, but those are two methods of one class on the same TheDebug
    global; this one crosses TextStream to Debug.
  * rb3-xenon leverage does not transfer. Its 483 template_sibling sites are
    only 17 MakeString; the bulk is STL element-type mismatches
    (__uninitialized_copy 81, _Destroy_Range 48, _M_find 43, DeleteAll 36) --
    structural, not an argument respelling.

Tool defect found in passing: namecheck_triage.py's LOCAL_GUARD rule ends the
enclosing function at @(\d+)$, so ...@Z@5 and ...@Z@4IA split differently and
two statics in the SAME function compare as different functions (3 pairs / 9
sites here). It should terminate at the mangling's own @Z/XZ. Not fixed here --
that file is another lane's.

Measured against main, pinned objdiff-cli-B, alias map regenerated on both
sides, one -o per arm, sidecars purged:

  none        43.730507% (4,973,656 B)  ->  43.730507% (4,973,656 B)   +0  -0
  name_check  42.635746% (4,849,144 B)  ->  42.714455% (4,858,096 B)   +20 -0

`none` is byte-identical and none_guard's 28,677-name ??_C@ fingerprint never
moved, across all four wave checkpoints. Exposed functions 320 -> 300.

+++ src/system/hamobj/SongLayout.cpp
@@ -183,7 +183,7 @@ void SongLayout::DumpPatterns() const {
             it->mInitialMeasureRange.start,
             it->mInitialMeasureRange.end
         );
-        MILO_LOG(" size=%d moves=%d\n", (int)it->mElements.size(), it->mNumMoves);
+        MILO_LOG(" size=%d moves=%d\n", it->mElements.size(), it->mNumMoves);
         i++;
         for (int j = 0; j < it->mNumMoves; j++) {
             MILO_LOG("   %s\n", it->mMoveParents[j]->Name());
